# source/app/blueprints/rest/openapi.generated.yaml
openapi: 3.1.0
info:
  title: IRIS
  version: 2.1.0
  description: "IRIS REST API \u2014 automatically generated from the Flask backend."
  contact:
    name: DFIR-IRIS
    email: contact@dfir-iris.org
  license:
    name: LGPLv3
    url: https://www.gnu.org/licenses/lgpl-3.0.txt
security:
- ApiKey: []
paths:
...
  /api/v2/manage/modules:
    get:
      operationId: list_modules_get
      summary: List modules
      tags:
      - ManageModules
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: add_module_post
      summary: Register a module
      tags:
      - ManageModules
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v2/manage/modules/{identifier}:
    delete:
      operationId: delete_module_delete
      summary: Delete a module
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: get_module_get
      summary: Get a module
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/{identifier}/disable:
    post:
      operationId: disable_module_post
      summary: Disable a module
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/{identifier}/enable:
    post:
      operationId: enable_module_post
      summary: Enable a module
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/{identifier}/export-config:
    get:
      operationId: export_module_config_get
      summary: Export a module configuration
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/{identifier}/import-config:
    post:
      operationId: import_module_config_post
      summary: Import a module configuration
      tags:
      - ManageModules
      security:
      - ApiKey: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/{identifier}/parameters/{param_name}:
    put:
      operationId: set_module_parameter_put
      summary: Update a module parameter
      tags:
      - ManageModules
      security:
      - ApiKey: []
      description: "The legacy API encoded `mod_id##param_name` as base64 in the URL \u2014\nwe don't\
        \ carry that scheme over: the v2 URL puts `module_id` and\n`param_name` as discrete path segments.\
        \ `path:` matches slashes so\nparameter names that contain `/` (rare but legal) round-trip cleanly."
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: integer
      - name: param_name
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/manage/modules/hooks:
    get:
      operationId: list_modules_hooks_get
      summary: List module hooks
      tags:
      - ManageModules
      security:
      - ApiKey: []
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer <api-key> \u2014 the key from your profile."
  responses:
    ApiError:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              data: {}
            required:
            - message
    Forbidden:
      description: Insufficient permissions to access resource
    NotFound:
      description: Resource not found
    Deleted:
      description: Resource successfully deleted
...